Add double check

+++ build/sync_check.sh
@@ -3,6 +3,11 @@
 LIGHT_PID=$(ps -ef | grep "syncmode light" | grep -v "grep" | awk '{print $1}')
 curl -s -X POST -H "Content-Type: application/json" --data '{"jsonrpc":"2.0","method":"eth_syncing","params":[],"id":1}'  http://localhost:8645 | grep 'false'
 if [ $? = 0 ];then
+    sleep 60;
+    curl -s -X POST -H "Content-Type: application/json" --data '{"jsonrpc":"2.0","method":"eth_syncing","params":[],"id":1}'  http://localhost:8645 | grep 'false'
+    if [ $? = 1 ];then
+        exit 0;
+    fi
     if [ ! -z "$LIGHT_PID" ]; then
         kill $LIGHT_PID
     fi
